requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
numpy>=1.23.0
plotly>=5.13.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
openpyxl>=3.1.0  # for Excel file support
pyarrow>=12.0.0  # for parquet file support
h5py>=3.8.0  # for HDF5 file support